itoa
Trait Integer

pub trait Integer: Sealed { }

An integer that can be written into an itoa::Buffer.

This trait is sealed and cannot be implemented for types outside of itoa.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.