Trait itoa::Integer

pub trait Integer: Sealed { }

An integer that can be written into an itoa::Buffer.

This trait is sealed and cannot be implemented for types outside of itoa.

Object Safety

This trait is not object safe.