Trait itoa::Integer

pub trait Integer: Sealed {
    const MAX_STR_LEN: usize;
}

An integer that can be written into an itoa::Buffer.

This trait is sealed and cannot be implemented for types outside of itoa.

Required Associated Constants

const MAX_STR_LEN: usize

The maximum length of string that formatting an integer of this type can produce on the current target platform.

Object Safety

This trait is not object safe.